api/api_requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
redis==5.0.1
PyJWT==2.8.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
Pillow==10.0.0
reportlab==4.0.7
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23